+++ src/Support/Source/Contracts/ObjectPayloadExtractor.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tabuna\Map\Support\Source\Contracts;
+
+interface ObjectPayloadExtractor
+{
+    /**
+     * Try extracting array payload from object source.
+     *
+     * @return array|null
+     */
+    public function extract(object $source): ?array;
+}